修改销售出库单同步金蝶加折扣价格

+++ erp-server/erp-server-wms/src/main/java/com/erp/server/wms/kingdee/impl/SyncKingdeeSoOutstockServiceImpl.java
@@ -212,7 +212,7 @@ public class SyncKingdeeSoOutstockServiceImpl implements SyncKingdeeSoOutstockSe
             resultMap.put("salesOrgCode", salesOrgCode);
         }
         if (soInfoById.getDiscountAmount() != null) {
-            resultMap.put("FAllDisCount", soInfoById.getDiscountAmount().intValue());
+            resultMap.put("FAllDisCount", soInfoById.getDiscountAmount());
         }
 
         List<String> soDetailIds = soDetailEntitieList.stream().map(SoDetailEntity::getId).collect(Collectors.toList());
